fix: fix intl for about
